first useable version!
* setup particle previewer

* still some weirdness with how it schedules the particle preview timer

* bug with particle list name, it reads it from a different row for some reason when you create a new particle, but when you refresh the list it picks up the correct name, gotta dig into it

* sliders are kind of hard to use since the range of values is so big, need to try and fix this somehow.

one idea is to set the sliders to a reasonable range (like -50 to 50) then after you take your finger off the slider, add the slider value to the real value and then set the slider back to 0, so you would just keep using the same slider over and over again to get to the value you wanted. it would be slower making huge changes but it would be a lot better making fine tune adjustments, which i think is a good trade off.

* need a randomize button

* need better default values

* for some reason the color variance makes things rainbow. i would want every particle to be the same random color, not every particle a different random color